Improve hit test methods (#6)
Hit testing methods were failing for views with gesture recognizers and for views that only responded to a specific type of event (example: dragging). I'm updating the screenHitPoint to use other methods for detecting if a view is hittable.

Additionally, I've also changed the default tap location behavior to return the main view instead of the center point of the screen. By using the view directly we can check if it's actually visible or hittable, resulting in significantly better error messages.